Week01-Historical Options Surface/requirements.txt
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
python_dateutil==2.8.2
scipy==1.6.2
tqdm==4.59.0
# yahoo_fin==0.8.9.1
